+++ src/containers/thread/DashboardThread/Layout/KanbanLayout/BgColorsSetter/KanbanList.tsx
