+++ src/containers/thread/DashboardThread/styles/layout/kanban_layout/bg_colors_setter/kanben_item.ts
